requirements.txt
bitstring
pyserial
psycopg2
SQLAlchemy